// src/store/recipe/recipe.types.ts
export enum RECIPES_ACTION_TYPES {
   FETCH_RECIPE_START = 'recipe/FETCH_RECIPE_START',
   FETCH_RECIPE_SUCCESS = 'recipe/FETCH_RECIPE_SUCCESS',
   FETCH_RECIPE_FAILED = 'recipe/FETCH_RECIPE_FAILED',
};

export enum RECIPES_SEARCH_ACTION_TYPES {
   FETCH_RECIPE_SEARCH_START = 'recipe/FETCH_RECIPE_SEARCH_START',
   FETCH_RECIPE_SEARCH_SUCCESS = 'recipe/FETCH_RECIPE_SEARCH_SUCCESS',
   FETCH_RECIPE_SEARCH_FAILED = 'recipe/FETCH_RECIPE_SEARCH_FAILED',
};

export enum RECIPES_FEATURED_ACTION_TYPES {
   FETCH_RECIPES_FEATURED_START = 'recipe/FETCH_RECIPES_FEATURED_START',
   FETCH_RECIPES_FEATURED_SUCCESS = 'recipe/FETCH_RECIPES_FEATURED_SUCCESS',
   FETCH_RECIPES_FEATURED_FAILED = 'recipe/FETCH_RECIPES_FEATURED_FAILED',
}

export enum RECIPE_BYID_ACTION_TYPES {
   FETCH_RECIPE_BYID_START = 'recipe/FETCH_RECIPE_BYID_START',
   FETCH_RECIPE_BYID_SUCCESS = 'recipe/FETCH_RECIPE_BYID_SUCCESS',
   FETCH_RECIPE_BYID_FAILED = 'recipe/FETCH_RECIPE_BYID_FAILED',
}